+++ lib/Comment.dart
@@ -1,8 +1,8 @@
-
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:flutter/material.dart';
 import 'package:final_project/TweetPage.dart';
+import 'package:flutter/material.dart';
 
+final commentsRef = FirebaseFirestore.instance.collection('comments');
 class Comment {
   String? id;
   final String userLongName;
@@ -59,6 +59,72 @@ class CommentsList extends StatelessWidget {
   Widget build(BuildContext context) {
     return Column(
       children: comments.map((comment) => CommentItem(comment: comment)).toList(),
+    );
+  }
+}
+
+class CreateComment extends StatefulWidget 
+{
+  const CreateComment({super.key});
+  @override
+  _CreateCommentState createState() => _CreateCommentState();
+}
+
+class _CreateCommentState extends State<CreateComment> 
+{
+  final userLongName = TextEditingController();
+  final userShortName = TextEditingController();
+  final text = TextEditingController();
+  final imageURL = TextEditingController();
+
+  void submitComment() 
+  {
+    final newComment = Comment(
+      userLongName: userLongName.text,
+      userShortName: userShortName.text,
+      timestamp: DateTime.now(),
+      text: text.text,
+      imageURL: imageURL.text,
+    );
+
+    Navigator.pop(context, newComment);
+  }
+
+  @override
+  Widget build(BuildContext context) 
+  {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text("Create A Comment"),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          children: [
+            TextField(
+              controller: userLongName,
+              decoration: const InputDecoration(hintText: 'Name'),
+            ),
+            TextField(
+              controller: userShortName,
+              decoration: const InputDecoration(hintText: 'Username'),
+            ),
+            TextField(
+              controller: text,
+              decoration: const InputDecoration(hintText: "Add Reply..."),
+            ),
+            TextField(
+              controller: imageURL, 
+              decoration: const InputDecoration(hintText: "Image URL (optional)"),
+            ),
+            const SizedBox(height: 20),
+            ElevatedButton(
+              onPressed: submitComment,
+              child: const Text("Comment"),
+            ),
+          ],
+        ),
+      ),
     );
   }
 }
